added delete routefor admin
