+++ client/app/AppRoutes.js
@@ -10,8 +10,9 @@ import Checkout from "../features/checkout/Checkout";
 import AdminPage from "../features/admin/AdminPage";
 import Cart from "../features/cart/Cart";
 import UpdateProduct from "../features/admin/UpdateProduct";
-import PageNotFound from "../features/PageNotFound";
 import AddProduct from "../features/admin/AddProduct";
+import MyAccount  from "../features/account/MyAccount"
+import PageNotFound from "../features/PageNotFound";
 
 /**
  * COMPONENT
@@ -35,6 +36,9 @@ const AppRoutes = () => {
             <Route to="/*" element={<MainPage />} />
             <Route path="/" element={<MainPage />} />
 
+            {/* My Account Page */}
+            <Route path='/my-account' element={<MyAccount/>}/>
+            
             {/* Product Routes */}
             <Route path="/products/:id" element={<SingleProductView />} />
             <Route path="/checkout" element={<Checkout />} />
