+++ client/features/account/MyAccount.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const MyAccount = () => {
+  return (
+    <div>MyAccount Page</div>
+  )
+}
+
+export default MyAccount
